requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.2.1
google-generativeai>=0.4.1
google-cloud-logging>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
